+++ lib/matrices.dart
@@ -2,4 +2,6 @@ library matrices;
 
 export 'src/matrix.dart';
 export 'src/square_matrix.dart';
+
+// Deprecated
 export 'src/emoji.dart';