+++ lib/src/emoji.dart
@@ -1,4 +1,5 @@
-/// 神奇的表情包方法
+/// Magic Emoji Methods(Deprecated)
+/// 神奇的表情包方法(已废弃)
 
 /// 快速生成0到n的向量
 List<double> o_n(int n) {